Move Spec, Suite, and CompositeTest to internal
