+++ src/main/java/com/greghaskins/spectrum/internal/Suite.java
@@ -1,10 +1,6 @@
-package com.greghaskins.spectrum;
+package com.greghaskins.spectrum.internal;
 
-import com.greghaskins.spectrum.internal.Child;
-import com.greghaskins.spectrum.internal.ConfiguredBlock;
-import com.greghaskins.spectrum.internal.NameSanitiser;
-import com.greghaskins.spectrum.internal.NotifyingBlock;
-import com.greghaskins.spectrum.internal.Parent;
+import com.greghaskins.spectrum.Block;
 import com.greghaskins.spectrum.model.BlockConfiguration;
 import com.greghaskins.spectrum.model.HookContext;
 import com.greghaskins.spectrum.model.Hooks;
@@ -18,7 +14,7 @@ import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 
-class Suite implements Parent, Child {
+public class Suite implements Parent, Child {
   private Hooks hooks = new Hooks();
 
   protected final List<Child> children = new ArrayList<>();
@@ -42,7 +38,7 @@ class Suite implements Parent, Child {
     void runChildren(final Suite suite, final RunNotifier notifier);
   }
 
-  static Suite rootSuite(final Description description) {
+  public static Suite rootSuite(final Description description) {
     return new Suite(description, Parent.NONE, Suite::defaultChildRunner,
         new TaggingFilterCriteria());
   }
@@ -66,7 +62,7 @@ class Suite implements Parent, Child {
     this.tagging = taggingFilterCriteria;
   }
 
-  Suite addSuite(final String name) {
+  public Suite addSuite(final String name) {
     return addSuite(name, Suite::defaultChildRunner);
   }
 
@@ -84,7 +80,7 @@ class Suite implements Parent, Child {
     return suite;
   }
 
-  Suite addCompositeSuite(final String name) {
+  public Suite addCompositeSuite(final String name) {
     final Suite suite =
         new CompositeTest(Description.createSuiteDescription(sanitise(name)), this,
             this.tagging.clone());
@@ -92,7 +88,7 @@ class Suite implements Parent, Child {
     return addedToThis(suite);
   }
 
-  Child addSpec(final String name, final Block block) {
+  public Child addSpec(final String name, final Block block) {
     final Child spec = createSpec(name, block);
     addChild(spec);
 
@@ -117,17 +113,17 @@ class Suite implements Parent, Child {
   }
 
   /**
-  * Adds a hook to be the first one executed before the block.
-  * This is the default. Hooks should be executed in the order they
-  * are declared in the test.
-  * @param hook to add
-  */
-  void addHook(final HookContext hook) {
+   * Adds a hook to be the first one executed before the block. This is the default. Hooks should be
+   * executed in the order they are declared in the test.
+   *
+   * @param hook to add
+   */
+  public void addHook(final HookContext hook) {
     this.hooks.add(hook);
   }
 
   private Hooks getHooksFor(final Child child) {
-    Hooks allHooks = parent.getInheritableHooks().plus(hooks);
+    Hooks allHooks = this.parent.getInheritableHooks().plus(this.hooks);
 
     return child.isAtomic() ? allHooks.forAtomic() : allHooks.forNonAtomic();
   }
@@ -138,7 +134,7 @@ class Suite implements Parent, Child {
     // all other hooks would be executed at suite level only - either for
     // each child of the suite, or once
 
-    return parent.getInheritableHooks().plus(hooks.forAtomic());
+    return this.parent.getInheritableHooks().plus(this.hooks.forAtomic());
   }
 
   /**
@@ -146,7 +142,7 @@ class Suite implements Parent, Child {
    *
    * @param tags required tags - suites must have at least one of these if any are specified
    */
-  void includeTags(final String... tags) {
+  public void includeTags(final String... tags) {
     this.tagging.include(tags);
   }
 
@@ -155,11 +151,11 @@ class Suite implements Parent, Child {
    *
    * @param tags excluded tags - suites and specs must not have any of these if any are specified
    */
-  void excludeTags(final String... tags) {
+  public void excludeTags(final String... tags) {
     this.tagging.exclude(tags);
   }
 
-  void applyPreconditions(Block block) {
+  public void applyPreconditions(Block block) {
     this.preconditions = Child.findApplicablePreconditions(block);
     applyPreconditions(block, this.tagging);
   }
@@ -203,8 +199,8 @@ class Suite implements Parent, Child {
     if (isEffectivelyIgnored()) {
       runChildren(notifier);
     } else {
-      hooks.once().sorted()
-          .runAround(description, notifier, () -> runChildren(notifier));
+      this.hooks.once().sorted()
+          .runAround(this.description, notifier, () -> runChildren(notifier));
     }
   }
 
@@ -219,7 +215,7 @@ class Suite implements Parent, Child {
     } else if (childIsNotInFocus(child)) {
       notifier.fireTestIgnored(child.getDescription());
     } else {
-      hooks.forThisLevel().sorted().runAround(child.getDescription(), notifier,
+      this.hooks.forThisLevel().sorted().runAround(child.getDescription(), notifier,
           () -> runChildWithHooks(child, notifier));
     }
   }
@@ -246,7 +242,7 @@ class Suite implements Parent, Child {
     return this.children.stream().mapToInt(Child::testCount).sum();
   }
 
-  void removeAllChildren() {
+  public void removeAllChildren() {
     this.children.clear();
   }
 
@@ -255,16 +251,16 @@ class Suite implements Parent, Child {
   }
 
   private String sanitise(final String name) {
-    return nameSanitiser.sanitise(name);
+    return this.nameSanitiser.sanitise(name);
   }
 
   @Override
   public boolean isEffectivelyIgnored() {
-    return ignored || !hasANonIgnoredChild();
+    return this.ignored || !hasANonIgnoredChild();
   }
 
   private boolean hasANonIgnoredChild() {
-    return children.stream()
+    return this.children.stream()
         .filter(child -> !child.isEffectivelyIgnored())
         .findFirst()
         .isPresent();
